refactor: split ExitProcessor.Core into behavior-themed submodules
fix: dialyzer

fix: groom and fix dialyzer warnings in ExitProcessor


refactor: split ExitProcessor.Core into behavior-themed submodules
